4.9.7.0 Added PcmData12 property to support 12bit PCM data for C140 chip. *Warning* For 12-bit PCM data, the Gain value should be lower than for 8-bit PCM data.

+++ src/devices/sound/c140.h
@@ -17,7 +17,7 @@
 // ======================> c140_device
  // Verified from schematics (24 bit address, 12(16? for C219) bit data)
 
-typedef int8_t (*C140_CALLBACK)(u8 pn, int32_t pos);
+typedef s32 (*C140_CALLBACK)(u8 pn, int32_t pos);
 
 class c140_device : public device_t,
 					public device_sound_interface,
